[Refactor]: 코드 리뷰 제안 사항 적용 및 무의미한 package 구조 재구성
