+++ hanbatbox-api/src/main/resources/application.yml
@@ -1,3 +1,6 @@
+server:
+  servlet:
+    context-path: /api
 spring:
   config:
     import: optional:file:.env[.properties]
