+++ hanbatbox-api/src/test/java/com/hbu/hanbatbox/S3/controller/S3ControllerTest.java
@@ -1,4 +1,4 @@
-package com.hbu.hanbatbox.module.S3.controller;
+package com.hbu.hanbatbox.S3.controller;
 
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doThrow;
@@ -6,7 +6,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.hbu.hanbatbox.module.S3.service.S3Service;
+import com.hbu.hanbatbox.S3.service.S3Service;
 import java.nio.file.Files;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -37,9 +37,9 @@ public class S3ControllerTest {
 
   @Test
   public void testUploadFileWithPassword() throws Exception {
-    String title = "Test Title";
+    String title = "Test.txt";
     String password = "secret";
-    String objectKey = "test-" + System.currentTimeMillis() + "-" + title;
+    String objectKey = "12341234-Test.txt";
 
     // S3 서비스 메서드가 호출될 때의 동작 정의
     doNothing().when(s3Service)
@@ -47,29 +47,29 @@ public class S3ControllerTest {
 
     // 파일 업로드 요청 보내기
     mockMvc.perform(
-            multipart("/api/s3/upload").file(file).param("title", title).param("password", password))
+            multipart("/file/upload").file(file).param("title", title).param("password", password))
         .andExpect(status().isOk()).andExpect(content().string("File uploaded successfully!"));
   }
 
   @Test
   public void testUploadFileWithoutPassword() throws Exception {
-    String title = "Test Title";
-    String objectKey = "file-123-test.txt"; // 예시 키
+    String title = "Test.txt";
+    String objectKey = "12341234-Test.txt";
 
     // S3 서비스 메서드가 호출될 때의 동작 정의
     doNothing().when(s3Service)
         .uploadFile(objectKey, title, null, Files.createTempFile("upload-", "test.txt"));
 
     // 파일 업로드 요청 보내기
-    mockMvc.perform(multipart("/api/s3/upload").file(file).param("title", title))
+    mockMvc.perform(multipart("/file/upload").file(file).param("title", title))
         .andExpect(status().isOk()).andExpect(content().string("File uploaded successfully!"));
   }
 
   @Test
   public void testUploadFileFailure() throws Exception {
-    String title = "Test Title";
+    String title = "Test.txt";
     String password = "secret";
-    String objectKey = "file-123-test.txt"; // 예시 키
+    String objectKey = "12341234-Test.txt";
 
     // S3 서비스 메서드가 실패할 때의 동작 정의
     doThrow(new RuntimeException("File upload failed!")).when(s3Service)
@@ -77,7 +77,7 @@ public class S3ControllerTest {
 
     // 파일 업로드 요청 보내기
     mockMvc.perform(
-            multipart("/api/s3/upload").file(file).param("title", title).param("password", password))
+            multipart("/file/upload").file(file).param("title", title).param("password", password))
         .andExpect(status().isInternalServerError())
         .andExpect(content().string("File upload failed!"));
   }
